django/0914/0914실습/requirements.txt
asgiref==3.2.10
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-appconf==1.0.4
django-extensions==3.0.5
django-imagekit==4.0.2
entrypoints==0.3
Faker==4.1.1
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
mysql-connector-python==8.0.21
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
